refactor: update namespace of the node
refactor: update namespace of the node

refactor: update namespace and directory structure for detection_by_tracker node
