+++ perception/detection_by_tracker/CMakeLists.txt
@@ -17,21 +17,15 @@ find_package(eigen3_cmake_module REQUIRED)
 find_package(Eigen3 REQUIRED)
 
 include_directories(
-  include
   SYSTEM
     ${EIGEN3_INCLUDE_DIRS}
     ${PCL_COMMON_INCLUDE_DIRS}
     ${PCL_INCLUDE_DIRS}
 )
 
-# Generate exe file
-set(DETECTION_BY_TRACKER_SRC
-  src/detection_by_tracker_node.cpp
-  src/utils/utils.cpp
-)
-
 ament_auto_add_library(${PROJECT_NAME} SHARED
-  ${DETECTION_BY_TRACKER_SRC}
+  src/node/detection_by_tracker_node.cpp
+  src/tracker/tracker_handler.cpp
 )
 
 target_link_libraries(${PROJECT_NAME}
@@ -43,9 +37,7 @@ rclcpp_components_register_node(${PROJECT_NAME}
   PLUGIN "autoware::detection_by_tracker::DetectionByTracker"
   EXECUTABLE detection_by_tracker_node
 )
-
-ament_auto_package(
-  INSTALL_TO_SHARE
+ament_auto_package(INSTALL_TO_SHARE
   launch
   config
 )
