+++ perception/detection_by_tracker/src/tool/debugger.hpp
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef DEBUGGER__DEBUGGER_HPP_
-#define DEBUGGER__DEBUGGER_HPP_
+#ifndef TOOL__DEBUGGER_HPP_
+#define TOOL__DEBUGGER_HPP_
 
-#include "autoware/shape_estimation/shape_estimator.hpp"
 #include "autoware/universe_utils/ros/debug_publisher.hpp"
 #include "autoware/universe_utils/system/stop_watch.hpp"
 #include "euclidean_cluster/euclidean_cluster.hpp"
@@ -46,7 +45,6 @@
 #include <deque>
 #include <memory>
 #include <vector>
-
 namespace autoware::detection_by_tracker
 {
 class Debugger
@@ -122,4 +120,4 @@ private:
 };
 }  // namespace autoware::detection_by_tracker
 
-#endif  // DEBUGGER__DEBUGGER_HPP_
+#endif  // TOOL__DEBUGGER_HPP_
