+++ perception/detection_by_tracker/src/tool/utils.hpp
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "utils.hpp"
+#ifndef TOOL__UTILS_HPP_
+#define TOOL__UTILS_HPP_
 
-#include <autoware_perception_msgs/msg/object_classification.hpp>
+#include "autoware_perception_msgs/msg/object_classification.hpp"
+
+#include <cstdint>
 
 namespace autoware::detection_by_tracker
 {
@@ -22,12 +25,26 @@ namespace utils
 {
 using Label = autoware_perception_msgs::msg::ObjectClassification;
 
-bool TrackerIgnoreLabel::isIgnore(const uint8_t label) const
+struct TrackerIgnoreLabel
 {
-  return (label == Label::UNKNOWN && UNKNOWN) || (label == Label::CAR && CAR) ||
-         (label == Label::TRUCK && TRUCK) || (label == Label::BUS && BUS) ||
-         (label == Label::TRAILER && TRAILER) || (label == Label::MOTORCYCLE && MOTORCYCLE) ||
-         (label == Label::BICYCLE && BICYCLE) || (label == Label::PEDESTRIAN && PEDESTRIAN);
-}
+  bool UNKNOWN;
+  bool CAR;
+  bool TRUCK;
+  bool BUS;
+  bool TRAILER;
+  bool MOTORCYCLE;
+  bool BICYCLE;
+  bool PEDESTRIAN;
+  bool isIgnore(const uint8_t label) const
+  {
+    return (label == Label::UNKNOWN && UNKNOWN) || (label == Label::CAR && CAR) ||
+           (label == Label::TRUCK && TRUCK) || (label == Label::BUS && BUS) ||
+           (label == Label::TRAILER && TRAILER) || (label == Label::MOTORCYCLE && MOTORCYCLE) ||
+           (label == Label::BICYCLE && BICYCLE) || (label == Label::PEDESTRIAN && PEDESTRIAN);
+  }
+};
+
 }  // namespace utils
 }  // namespace autoware::detection_by_tracker
+
+#endif  // TOOL__UTILS_HPP_
